Initial test_model_from.py
